[chore] Xavier v1.0 Release Blocker Cleanup
- Security Purge: Removed 'dev-token' defaults and XAVIER_DEV_MODE fallback from authentication logic, scripts, and documentation.
- Rust Core Stability: Fixed all clippy warnings and ensured doctests pass across the workspace.
- Storage Isolation: Verified workspace-level data isolation and added a dedicated integration test.
- Documentation Sync: Updated QUICKSTART.md with verified CLI commands and fixed formatting.
- Issue Tracking: Updated v1.0 release blocker markdown files to reflect completed status.

+++ src/onboarding/scanner.rs
@@ -349,7 +349,7 @@ fn detect_disk_free_gb() -> f64 {
             unsafe {
                 let mut stat: libc::statvfs = std::mem::zeroed();
                 if libc::statvfs(cstr.as_ptr(), &mut stat) == 0 {
-                    Some((stat.f_bsize as u64 * stat.f_bavail) as f64 / (1024.0 * 1024.0 * 1024.0))
+                    Some((stat.f_bsize * stat.f_bavail) as f64 / (1024.0 * 1024.0 * 1024.0))
                 } else {
                     None
                 }
